Move NodeBuilder to a dedicated file
